backend/requirements.txt
asgiref==3.8.1
Django==4.2
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
pillow==11.0.0
PyJWT==2.10.1
rest-framework-simplejwt==0.0.2
sqlparse==0.5.2
typing_extensions==4.12.2
uuid==1.30
